Added initial HTML and CSS. Vanilla CSS, HTML, and JS.
Will be used to create the UI and handle the interactivity required.

A basic UI has been created although it isn't currently interactive,
with the exception of the CSS styles that handle hover states.

A search input has been provided along with a search button to exec the
queries. An un-ordered list serves as the container for the search
histories and one <li> has been hardcoded to test the styles and serve
as a template. This item will be dynamically created for each search
via JavaScript.

I was going to use Tailwind CSS to style the UI, but I decided to see
how much it really saved over vanilla CSS. My first rendition used
Tailwinds and I have to say, I did this with way less CSS and my HTML
isn't a cluster F to read.